detail path

+++ src/main/java/com/ll/jsp/board/boundedContext/article/repository/ArticleRepository.java
@@ -39,7 +39,7 @@ public class ArticleRepository {
         return id;
     }
 
-    public Article findById(int id) {
+    public Article findById(long id) {
         return articleList.stream()
                 .filter(article -> article.getId() == id)
                 .findFirst()
